Prefer OpenSSL's SIPHASH implementation when available
OpenSSL library is significantly faster than the reference implementation
(almost 2x). Prefer using this when available. The API for using the SIPHASH
MAC is different enough from using normal HMAC or Digest that we already
implement that combining them into one API does not make sense.

SIPHASH is only available on OpenSSL 3.1 and later. We still check for
support on 3.0 and later as the whole API to allow using the SIPHASH alrady
exists in OpenSSL 3.0. Some of the later OpenSSL 3.0.x might get support
for it. Theoretically, a provider can be loaded in OpenSSL 3.0 that
implements SIPHASH.

Change-Id: I09aa27caa1a3aab0d1be6118b26d54a1c1bf7aa0